requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
gunicorn==21.2.0
